+++ src/calculations/calculations.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
+import { ConfigModule } from '@nestjs/config';
 import { CurrencyModule } from '../currency/currency.module';
 import { CustomsModule } from '../customs/customs.module';
 import { CalculationsService } from './calculations.service';
 
 @Module({
-  imports: [CurrencyModule, CustomsModule],
+  imports: [ConfigModule, CurrencyModule, CustomsModule],
   providers: [CalculationsService],
   exports: [CalculationsService],
 })