+++ src/currency/currency.module.ts
@@ -1,10 +1,12 @@
 import { Module } from '@nestjs/common';
+import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { YuanRate } from '../entities/yuan-rate.entity';
+import { PlaywrightModule } from '../common/playwright.module';
 import { CurrencyService } from './currency.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([YuanRate])],
+  imports: [ConfigModule, TypeOrmModule.forFeature([YuanRate]), PlaywrightModule],
   providers: [CurrencyService],
   exports: [CurrencyService],
 })